requirements.txt
agno
qdrant-client
openai
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
google-generativeai
Flask
